Building Class created

+++ MLAK-Clone/Map.cpp
@@ -9,7 +9,7 @@ Map::~Map()
 	delete tiles;
 }
 
-void Map::Create(int w, int h)
+void Map::Create(size_t w, size_t h)
 {
 	tiles = new Tile[w * h];
 }
